fixed deleting

+++ src/main/java/com/coinsearch/service/CoinCapService.java
@@ -1,8 +1,12 @@
 package com.coinsearch.service;
 import com.coinsearch.exception.EntityNotFoundException;
+import com.coinsearch.model.Chain;
 import com.coinsearch.model.CryptoData;
+import com.coinsearch.model.Person;
 import com.coinsearch.model.CryptocurrencyData;
+import com.coinsearch.repository.ChainRepository;
 import com.coinsearch.repository.CryptocurrencyRepository;
+import com.coinsearch.repository.PersonRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
@@ -14,11 +18,15 @@ public class CoinCapService {
 
     private final RestTemplate restTemplate;
     private final CryptocurrencyRepository cryptoRepository;
+    private final PersonRepository personRepository;
+    private final ChainRepository chainRepository;
     private static final String ERROR_MESSAGE = "Crypto does not exist with given id: ";
 
-    public CoinCapService(RestTemplate restTemplate, CryptocurrencyRepository cryptoRepository) {
+    public CoinCapService(RestTemplate restTemplate, CryptocurrencyRepository cryptoRepository, PersonRepository personRepository, ChainRepository chainRepository) {
         this.restTemplate = restTemplate;
         this.cryptoRepository = cryptoRepository;
+        this.personRepository = personRepository;
+        this.chainRepository = chainRepository;
     }
 
     public CryptoData createCryptocurrency(String cryptocurrency){
@@ -59,8 +67,39 @@ public class CoinCapService {
         CryptoData cryptoData = cryptoRepository.findById(Math.toIntExact(cryptoId)).orElseThrow(
                 () -> new EntityNotFoundException(ERROR_MESSAGE + cryptoId)
         );
-        if (cryptoData != null) {
+        if (cryptoData.getPersons().size() != 0){
+            throw new RuntimeException("Can't delete crypto " + cryptoId + " because people are using it. Try deleting this crypto from a specified person.");
+        }
+        if (cryptoData.getChain() != null){
+            throw new RuntimeException("Can't delete crypto " + cryptoId + " because it is in a chain. Try deleting this crypto from a specified chain.");
+        }
+        else if (cryptoData != null) {
             cryptoRepository.deleteById(Math.toIntExact(cryptoId));
         }
+    }
+
+    public void deleteCryptoFromPerson(Long cryptoId, Long personId) {
+        CryptoData cryptoData = cryptoRepository.findById(Math.toIntExact(cryptoId)).orElseThrow(
+                () -> new EntityNotFoundException(ERROR_MESSAGE + cryptoId)
+        );
+        Person person = personRepository.findById(Math.toIntExact(personId)).orElseThrow(
+                () -> new EntityNotFoundException("Person does not exist with given id: " + personId)
+        );
+
+        person.getCryptocurrencies().remove(cryptoData);
+        personRepository.save(person);
+    }
+
+    public void deleteCryptoFromChain(Long cryptoId, Long chainId) {
+        CryptoData cryptoData = cryptoRepository.findById(Math.toIntExact(cryptoId)).orElseThrow(
+                () -> new EntityNotFoundException(ERROR_MESSAGE + cryptoId)
+        );
+        Chain chain = chainRepository.findById(Math.toIntExact(chainId)).orElseThrow(
+                () -> new EntityNotFoundException("Chain does not exist with given id: " + chainId)
+        );
+
+        chain.getCryptocurrencies().remove(cryptoData);
+        cryptoData.setChain(null);
+        chainRepository.save(chain);
     }
 }
